libs/common_config/pyproject.toml
[project]
name = "common-config"
version = "0.1.0"
dependencies = [
    "python-dotenv",
    "pydantic-settings",
]
